+++ websiteapp/const.py
@@ -20,8 +20,8 @@ from typing import List
 
 class Version(Enum):
     MAJOR = 1
-    MINOR = 3
-    PATCH = 1
+    MINOR = 4
+    PATCH = 0
 
     @classmethod
     def as_string(cls) -> str: